holes: encode and decode sparse files
In "encode" mode, reads a file or stdin stream that may or may not
be sparse, finding the areas that "look sparse" (are all-zero-
bytes), and writes an output format (to file or stdout) that run-
length-encodes "long" sparse areas.

In "decode" mode, reads a file or stdin stream with this kind
of run-length-encoding and writes to an output stream with seeks
and ftruncate() calls as needed to create holes.

Note that this can create holes that were not originally present.