dotfiles: manipulate home dir .foorc etc
The dotfiles program is meant for manipulating ~/.foorc
files and the like, while keeping the source of those
files elsewhere, such as in this repository.

Checkpoint this version, before adding fancier options and
"going live".  The immediate plan is to add the ability to
save existing .foorc files in a tarball, and restore from
the tarball (essentially reversing the installation from the
Dotfiles directory).

(Due to .vim and pathogen and the like, I may also add an
option to indirect to Git repositories.)