dotfiles: reshuffle work-to-do code
Make a new worklist object that accumulates the lists
of work to do.  This is saved purely by pathname, so we
get to release the FileInfo objects sooner.

Removes the need for blank_finfo() entirely, and makes
the main worker function clearer.

(Also changes __str__ formatting of FileInfo to use full
path; this has only been used during debug.)