dotfiles: add --tar option
Using --tar, we save all the original dot-files to a tarball.

This still needs a way to reinstall the original tarball in
place of the installed Dotfiles, but at this point we are in
pretty good shape, since simply extracting the tar file over
the installed files is usually sufficient.