bashrc: fix untracked file detection
The `git ls-files` operation must be run in the top level of
the Git repository, as there is no option to say "all files"
and the default is to start at the current sub-directory of
the work-tree.

The options here were to use --show-toplevel or --show-cdup.
The --show-cdup method seems nicer to me, but it has a bit of
a glitch: it prints the empty string if you are already at the
top level.  Fortunately calling git -C "" works OK (as tested on
BSD and Linux at least).