add script to rerun smudge filters
Git should provide, as an environment variable, the target HEAD
when doing "git checkout <branchname>" or "git checkout <commit>"
so that this can be passed to "git describe".

It doesn't, so here is a Python script to let one re-run smudge
filters, post-checkout.

See https://stackoverflow.com/q/49667015/1256452 for more about
this.