vimrc: add \c for toggling column highlights
The highlighting is useful for the Go programming I've
been doing, but makes cut-and-paste go wrong as vim
writes lots of blanks to get to the highlight columns.
Add a function and bind it to \c for toggling the
highlighting.

Note: it's necessary to ^L-redraw before cutting and
pasting.  I tried putting the ^L into the binding, but
for some reason that only eliminates *some* of the
extra blanks, not all of them.