Dotfiles/profile: avoid bash aliases
Bash does not expand aliases in non-interactive scripts.
When such a script reads .bashrc, the shell-is-bash alias
gets ignored and the test to see which dot-file to use
internally fails.  Use a shell function instead.

(There is an option to make bash expand aliases even in
scripts, but I went the other way here.)