fbsd-r2g: improve function
The FreeBSD Git mirror now sometimes contains notes about
revisions that never make it into the Git mirror, so that
fbsd-r2g finds commit hashes that do not exist.  Make the
script check for these cases.

Add comments to describe the internal operation, too, as
it was a bit puzzling after a lot of years away from this
script.